DWIMy filterspecs for dotest.py
dotest.py currently requires a filterspec to be of the
form `TestCase.test_method`.   This patch makes it more
flexible, so you can pass TestModule.TestCase.test_method
or TestModule.TestCase or TestCase.test_method or just
test_method.

This makes it more convenient to just copy a test name
out of the terminal after running a bunch of tests and use
it as a filterspec.