Fix test breakage caused by r374424
The build directory name is based on the test method name, so having
two test methods with the same name in the same test file is a
problem, even if they're in different test classes.

The problem is fixed just by changing the name of one of the test
methods.