[LLDB][PythonFile] fix dangerous borrow semantics on python2
Summary:

It is, I think, inherently unsafe to hand a borrowed native file
object to a python program on python 2.   This is because the python
file object must be created with a FILE*, not a descriptor.     Even
holding on to such a reference past its legal scope can result in
use-after-free violations.   Pytyhon programs are not prepared to deal
with that kind of requirement.

Python 3 does not suffer from this issue.

Reviewers: labath, mgorny

Subscribers: lldb-commits

Tags: #lldb

Differential Revision: https://reviews.llvm.org/D69532